new files router/event.py and schemas/event.py
